+++ internal/common/error_message.go
@@ -1,0 +1,9 @@
+package errmessage
+
+const (
+	GET_TAG_ERROR     = "Error ocurred fetching tag information"
+	CREATE_TAG_ERROR  = "Error occurred while creating tag information"
+	UPDATE_TAG_ERROR  = "Error occurred while updating tag information"
+	DELETE_TAG_ERROR  = "Error occurred while deleting tag information"
+	JSON_DECODE_ERROR = "Error ocurred while decoding request body"
+)
