+++ internal/handler/taghandler.go
@@ -5,50 +5,29 @@ import (
 	"net/http"
 	"strconv"
 
+	errmessage "github.com/AnishriM/expenses-diary/internal/common"
 	"github.com/AnishriM/expenses-diary/internal/response"
 	"github.com/AnishriM/expenses-diary/internal/services/tag"
 	"github.com/gorilla/mux"
 )
 
-type Handler struct {
-	Router *mux.Router
-	DB     *tag.DBService
-}
-
-func NewHandler(db *tag.DBService) *Handler {
-	return &Handler{
-		Router: mux.NewRouter(),
-		DB:     db,
-	}
-}
-func (h *Handler) SetupRoutes() error {
-	println("Setting up Routes")
-	h.Router.HandleFunc("/api/tag/{id}", h.GetTagByID).Methods("GET")
-	h.Router.HandleFunc("/api/tag", h.GetAllTags).Methods("GET")
-	h.Router.HandleFunc("/api/tag/{id}", h.UpdateTag).Methods("PUT")
-	h.Router.HandleFunc("/api/tag/{id}", h.DeleteTag).Methods("DELETE")
-	h.Router.HandleFunc("/api/tag", h.CreateTag).Methods("POST")
-	return nil
-}
-
 func (h *Handler) GetTagByID(w http.ResponseWriter, r *http.Request) {
-	println("hit GetTagByID endpoint")
 	vars := mux.Vars(r)
 	id := vars["id"]
 	var uid uint64
 	var err error
 	if uid, err = strconv.ParseUint(id, 10, 64); err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error ocurred while parsing the int",
-			Error:   err,
+			Message: errmessage.GET_TAG_ERROR,
+			Error:   err.Error(),
 		})
 		return
 	}
 	tag, err := tag.GetTagByID(h.DB, uint(uid))
 	if err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error ocurred getting tag",
-			Error:   err,
+			Message: errmessage.GET_TAG_ERROR,
+			Error:   err.Error(),
 		})
 		return
 	}
@@ -56,12 +35,11 @@ func (h *Handler) GetTagByID(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *Handler) GetAllTags(w http.ResponseWriter, r *http.Request) {
-	println("hit GetAllTags endpoint")
 	tags, err := tag.GetAllTags(h.DB)
 	if err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error ocurred getting tags",
-			Error:   err,
+			Message: errmessage.GET_TAG_ERROR,
+			Error:   err.Error(),
 		})
 		return
 	}
@@ -69,7 +47,6 @@ func (h *Handler) GetAllTags(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *Handler) UpdateTag(w http.ResponseWriter, r *http.Request) {
-	println("hit UpdateTag endpoint")
 	vars := mux.Vars(r)
 	id := vars["id"]
 	var uid uint64
@@ -77,8 +54,8 @@ func (h *Handler) UpdateTag(w http.ResponseWriter, r *http.Request) {
 
 	if uid, err = strconv.ParseUint(id, 10, 64); err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error ocurred getting tags",
-			Error:   err,
+			Message: errmessage.GET_TAG_ERROR,
+			Error:   err.Error(),
 		})
 		return
 	}
@@ -86,8 +63,8 @@ func (h *Handler) UpdateTag(w http.ResponseWriter, r *http.Request) {
 	var newTag tag.Tag
 	if err := json.NewDecoder(r.Body).Decode(&newTag); err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error ocurred decoding body",
-			Error:   err,
+			Message: errmessage.JSON_DECODE_ERROR,
+			Error:   err.Error(),
 		})
 		return
 	}
@@ -95,8 +72,8 @@ func (h *Handler) UpdateTag(w http.ResponseWriter, r *http.Request) {
 	tag, err := tag.UpdateTag(uint(uid), newTag.Name, h.DB)
 	if err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error ocurred updating tags",
-			Error:   err,
+			Message: errmessage.UPDATE_TAG_ERROR,
+			Error:   err.Error(),
 		})
 		return
 	}
@@ -104,7 +81,6 @@ func (h *Handler) UpdateTag(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *Handler) DeleteTag(w http.ResponseWriter, r *http.Request) {
-	println("hit DeleteTag endpoint")
 	vars := mux.Vars(r)
 	id := vars["id"]
 	var uid uint64
@@ -112,8 +88,8 @@ func (h *Handler) DeleteTag(w http.ResponseWriter, r *http.Request) {
 
 	if uid, err = strconv.ParseUint(id, 10, 64); err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error ocurred deleting tag",
-			Error:   err,
+			Message: errmessage.DELETE_TAG_ERROR,
+			Error:   err.Error(),
 		})
 		return
 	}
@@ -121,8 +97,8 @@ func (h *Handler) DeleteTag(w http.ResponseWriter, r *http.Request) {
 	tag, err := tag.DeleteTag(uint(uid), h.DB)
 	if err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error ocurred updating tags",
-			Error:   err,
+			Message: errmessage.DELETE_TAG_ERROR,
+			Error:   err.Error(),
 		})
 		return
 	}
@@ -130,23 +106,21 @@ func (h *Handler) DeleteTag(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *Handler) CreateTag(w http.ResponseWriter, r *http.Request) {
-	println("hit CreateTag endpoint")
 	var err error
 	var newtag tag.Tag
 
 	if err := json.NewDecoder(r.Body).Decode(&newtag); err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error occured whil decoding json body",
-			Error:   err,
+			Message: errmessage.JSON_DECODE_ERROR,
+			Error:   err.Error(),
 		})
 	}
-	println("Tag Name:" + newtag.Name)
 
 	newtag, err = tag.CreateTag(newtag.Name, h.DB)
 	if err != nil {
 		response.SendErrorResponse(w, response.Response{
-			Message: "Error ocurred updating tags",
-			Error:   err,
+			Message: errmessage.CREATE_TAG_ERROR,
+			Error:   err.Error(),
 		})
 		return
 	}
